# backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: https://localhost:8443
    description: Generated server url
...
paths:
  /api/v1/profile/{id}:
    get:
      tags:
        - Profile API
      summary: Get profile
      description: Returns complete profile data for a given user id.
      operationId: profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ProfileDTO"
    put:
      tags:
        - Profile API
      summary: Update profile (JSON)
      description: Updates profile data from JSON payload for the authenticated owner.
      operationId: editProfileMultipart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                request:
                  $ref: "#/components/schemas/ProfileUpdateRequestDTO"
                imageFile:
                  type: string
                  format: binary
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
...
  /api/v1/statistics/lessons-learned:
    get:
      tags:
        - Statistics API
      summary: Lessons learned
      description: Returns monthly lessons completed by the user.
      operationId: lessonsLearned
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ChartDTO"
  /api/v1/statistics/created-course-status:
    get:
      tags:
        - Statistics API
      summary: Created course completion status
      description: Returns completion breakdown for students of a course.
      operationId: createdCourseStatus
      parameters:
        - name: courseId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ChartDTO"
  /api/v1/statistics/course-user-progress:
    get:
      tags:
        - Statistics API
      summary: User progress in course
      description: Returns completed vs pending lessons for one user in one course.
      operationId: courseUserProgress
      parameters:
        - name: courseId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ChartDTO"
  /api/v1/statistics/course-tags:
    get:
      tags:
        - Statistics API
      summary: Common course tags
      description: Returns overlap between user interests and course tags.
      operationId: courseTags
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: courseId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ChartDTO"
  /api/v1/statistics/course-progress:
    get:
      tags:
        - Statistics API
      summary: Course progress
      description: Returns completed vs in-progress course counts for a user.
      operationId: courseProgress
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ChartDTO"
  /api/v1/statistics/course-genders:
    get:
      tags:
        - Statistics API
      summary: Course genders
      description: Returns gender distribution for enrollments in a course.
      operationId: courseGenders
      parameters:
        - name: courseId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ChartDTO"
  /api/v1/statistics/course-ages:
    get:
      tags:
        - Statistics API
      summary: Course ages
      description: Returns age-range distribution for enrollments in a course.
      operationId: courseAges
      parameters:
        - name: courseId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/ChartDTO"
...
  /api/v1/auth/register/availability:
    get:
      tags:
        - Authentication API
      operationId: checkAvailability_1
      parameters:
        - name: username
          in: query
          required: false
          schema:
            type: string
        - name: email
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: boolean
...
components:
  schemas:
    ProfileUpdateRequestDTO:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        country:
          type: string
        shortDescription:
          type: string
        currentGoal:
          type: string
        weeklyRoutine:
          type: string
        comunity:
          type: string
...
    ChartDTO:
      type: object
      properties:
        chartTitle:
          type: string
        chartType:
          type: string
        chartLabels:
          type: array
          items:
            type: string
        chartValues:
          type: array
          items:
            type: number
            format: double
    ProfileDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        country:
          type: string
        shortDescription:
          type: string
        currentGoal:
          type: string
        weeklyRoutine:
          type: string
        comunity:
          type: string
        profileImage:
          type: string
        isProfileOwner:
          type: boolean
        userType:
          type: string
        completedCourses:
          type: integer
          format: int32
        completedCourseNames:
          type: array
          items:
            type: string
        inProgressCount:
          type: integer
          format: int32
        userTags:
          type: array
          items:
            type: string
          uniqueItems: true
        subscribedCourses:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
        userEvents:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
        averageProgress:
          type: integer
          format: int32
        totalEnrollments:
          type: integer
          format: int32
        totalLessonsCompleted:
          type: integer
          format: int64
        completedLessonsThisMonth:
          type: integer
          format: int64
        averageLessonsPerMonth:
          type: string
        createdCourses:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
        hasMultipleCourses:
          type: boolean
...