# backend/scam-g18/src/main/java/es/codeurjc/scam_g18/controller/RegisterWebController.java
package es.codeurjc.scam_g18.controller;

import java.io.IOException;
import java.util.HashMap;
import java.util.Map;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseBody;
import org.springframework.web.multipart.MultipartFile;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;

import es.codeurjc.scam_g18.service.EmailService;
import es.codeurjc.scam_g18.service.UserService;

@Controller
@Tag(name = "Registration Web API", description = "AJAX helper endpoints used in web registration")
public class RegisterWebController {

    @Autowired
    private EmailService emailService;

    @Autowired
    private UserService userService;

    @GetMapping("/register/check-availability")
    @Operation(summary = "Check availability (web)", description = "Checks whether username and email are already in use for web registration.")
    @ResponseBody
    public Map<String, Boolean> checkAvailability(
            @RequestParam(required = false) String username,
            @RequestParam(required = false) String email) {

        Map<String, Boolean> response = new HashMap<>();

        boolean usernameTaken = userService.usernameExists(username);
        boolean emailTaken = userService.emailExists(email);

        response.put("usernameTaken", usernameTaken);
        response.put("emailTaken", emailTaken);
        response.put("available", !usernameTaken && !emailTaken);

        return response;
    }

    @PostMapping("/register")
    public String registerUser(@RequestParam(required = false) String username,
            @RequestParam(required = false) String email,
            @RequestParam(required = false) String password, @RequestParam(required = false) String gender,
            @RequestParam(required = false) String birthDate,
            @RequestParam(required = false) String country,
            @RequestParam(value = "image", required = false) MultipartFile imageFile,
            jakarta.servlet.http.HttpServletRequest request,
            org.springframework.web.servlet.mvc.support.RedirectAttributes redirectAttributes)
            throws IOException, java.sql.SQLException {

        String validationErrors = userService.validateUserAttributes(username, email, password, birthDate, gender,
                country);
        if (validationErrors != null) {
            redirectAttributes.addFlashAttribute("error", validationErrors);
            return "redirect:/register";
        }

        boolean registered = userService.registerUser(username, email, password, gender, birthDate, country, imageFile);
        if (!registered) {
            redirectAttributes.addFlashAttribute("error", "El nombre de usuario o correo electrónico ya existen.");
            return "redirect:/register";
        }

        emailService.newAccountMessage(email, username);

        try {
            request.login(username, password);
        } catch (jakarta.servlet.ServletException e) {
            return "redirect:/login";
        }

        return "redirect:/profile/me";
    }
}